replace native type 'unsigned' to 'NSUInteger'

+++ fe-aqua/SGArray.m
@@ -13,55 +13,55 @@
 
 - (id) init
 {
-    items = NULL;
-    count = 0;
-    capacity = 0;
-    
-    return self;
+	items = NULL;
+	count = 0;
+	capacity = 0;
+	
+	return self;
 }
 
 - (void) dealloc
 {
-    free (items);
-    [super dealloc];
+	free (items);
+	[super dealloc];
 }
 
 - (NSUInteger) count
 {
-    return count;
+	return count;
 }
 
 - (id) objectAtIndex:(NSUInteger) index
 {
-    return (id) items [index];
+	return (id) items [index];
 }
 
 - (void) addObject:(id) object
 {
-    unsigned where = count ++;
-    
-    if (count > capacity)
-    {
-        capacity += 1 + capacity / 10;
-        items = (void **) realloc (items, capacity * sizeof (void *));
-    }
-    
-    items [where] = object;
+	NSUInteger where = count ++;
+	
+	if (count > capacity)
+	{
+		capacity += 1 + capacity / 10;
+		items = (void **) realloc (items, capacity * sizeof (void *));
+	}
+	
+	items [where] = object;
 }
 
 - (void) removeObject:(id) object
 {
-    unsigned s = 0;
-    unsigned d = 0;
-    
-    for ( ; s < count; s++)
-    {
-        if (items [s] == object)
-            continue;
-        items [d++] = items [s];
-    }
-    
-    count = d;
+	NSUInteger s = 0;
+	NSUInteger d = 0;
+	
+	for ( ; s < count; s++)
+	{
+		if (items [s] == object)
+			continue;
+		items [d++] = items [s];
+	}
+	
+	count = d;
 }
 
 @end
